requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
google-genai
motor
